Fix Imports, Naming, CSS Globals
