+++ src/components/notebooks/cards/notebook-plus.tsx
@@ -1,0 +1,8 @@
+
+export default function NotebookPlusCard() {
+  return (
+    <a href="#/" className="rounded-notebook-card justify-center min-w-[120px] max-w-[380px] bg-component-faded">
+      <h1 className="text-5xl text-center self-center text-faded" >+</h1>
+    </a>
+  )
+}
