--- src/components/notebooks/cards/notebookadd.tsx
@@ -1,9 +1,0 @@
-
-export default function NotebookAddCard() {
-    return (
-        <a href="#/" className="flex flex-col justify-center min-w-[120px] max-w-[380px] p-2 rounded-md 
-                                bg-component-faded hover:shadow-xl transition-shadow duration-300 ease-in-out">
-            <h1 className="text-5xl text-center self-center text-faded" >+</h1>
-        </a>
-    )
-}
