+++ src/styles/globals.css
@@ -185,3 +185,19 @@ input.invalid {
   background-color: var(--new-button-background-color);
   color: var(--new-button-text-color);
 }
+
+.rounded-notebook-card {
+  /* flex flex-col min-w-[120px] max-w-[380px] p-2 rounded-md 
+  bg-component hover:shadow-xl transition-shadow duration-300 ease-in-out */
+
+  display: flex;
+  flex-direction: column;
+	padding: 0.5rem;
+  border-radius: 0.375rem;
+}
+.rounded-notebook-card:hover {
+  box-shadow: 0 20px 25px -5px rgb(0 0 0 / 0.1), 0 8px 10px -6px rgb(0 0 0 / 0.1);
+  transition-property: box-shadow;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
+}
